This is the first commit
All the required files to run the login demo with servlet and jsp are the LoginServlet.java and index.jsp. the login avatar.png is just to include a decorative avatar image on the login page. You can just create a new project with any name of your choice to use